terraform/src/remove_owner_role/requirements.txt
google_api_python_client
google-auth
google-api-python-client